+++ src/ts/savedData.ts
@@ -1,4 +1,4 @@
-import { MapSetData } from './beatmap/map';
+import { BeatmapSetData } from './beatmap/map';
 import { Contributor } from './beatmap/contributor';
 import { BeatmapInfo } from './beatmap/info';
 import { BPMChange } from './beatmap/bpm';
@@ -6,7 +6,7 @@ import { BPMChange } from './beatmap/bpm';
 // TODO: structure bpm change for certain use
 interface SavedData {
     _mapInfo?: BeatmapInfo;
-    _mapSet?: MapSetData[];
+    _mapSet?: BeatmapSetData[];
     _contributors?: Contributor[];
     _analysis?: any;
     _duration?: number;